EP-3491 (#1338)
* EP-3491

* julkaisutestin korjausyritys

* julkaisuit test fix

* war -> jar

* rebase

* C:/Program Files/Git/api change

* cache fix

* deps fix

* rebase

* deps fixes

* tomcat comments

* https pakotus

* spring config security

* cas prop fix

* hibernate.id.new_generator_mappings

* cache cleans

* hibernate cache fix

* xml to java config

* springfox downgrade

* specs

* test fix

* test fix

* test fix

* rakenne haku fix

* logging

* log fix

* specs

* repo fix

* testfix

* ch.qos.logback fix

* cleanup

* old conf away

* cleanup

* interceptor fix

* cas client session

* redirect

* dev auth

* logout changes

* auditlogger pois

* delete auditlogger

* auditlogger back

* fixes

---------

+++ eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/repository/custom/TiedoteRepositoryImpl.java
@@ -48,7 +48,7 @@ public class TiedoteRepositoryImpl implements TiedoteRepositoryCustom {
     public Page<Tiedote> findBy(PageRequest page, TiedoteQuery tquery) {
         TypedQuery<Long> countQuery = getCountQuery(tquery);
         TypedQuery<Tuple> query = getQuery(page, tquery);
-        query.setFirstResult(page.getOffset());
+        query.setFirstResult(Long.valueOf(page.getOffset()).intValue());
         query.setMaxResults(page.getPageSize());
 
         log.debug(query.unwrap(Query.class).getQueryString());
